Tunning PID

+++ software/apps/combined_implementation/pid.h
@@ -25,11 +25,11 @@ extern float MIN_DUTY_CYCLE;
 //Takes the angle theta the bike makes with the vertical axis, return the motor
 //duty cycle required to correct the error using PID
 //Returns a number between 0 and 100 and a rotation direction
-void duty_cycle_PID(float theta, float time_stamp, uint8_t* duty_cyle, int8_t* direction);
+void duty_cycle_PID(float theta, float time_stamp, float* duty_cyle, int8_t* direction);
 
 //Takes and angle theta and outputs a simple proportionnal control duty cycle
 //between 0 and 100 and a direction
-void duty_cycle_proportionnal(float theta, uint8_t* duty_cyle, int8_t* direction);
+void duty_cycle_proportionnal(float theta, float* duty_cyle, int8_t* direction);
 
 //Resets the integral used in the PID. Takes the current clock time in clock cycles
 void reset_integral(int time_stamp);
